move the JWT dump function in to models.base
prevent circular includes
