+++ models/base.py
@@ -1,7 +1,25 @@
+from pydantic import BaseSettings
 from pydantic import BaseModel as PydBaseModel
 import jwt
 import os
-from tools.genjwts import dumps as jwt_dumps
+
+class Settings(BaseSettings):
+    jwt_algo: str = "HS256"
+    jwt_secret: str = "lolchangeme"
+
+settings = Settings()
+
+def jwt_dumps(value: 'IdentityPayload',
+          algo: str = settings.jwt_algo,
+          secret: str = settings.jwt_secret) -> str:
+    """
+    Serialize a JWT to a compact encoding
+    """
+    return jwt.encode(
+        value.dict(),
+        secret,
+        algo
+    )
 
 class BaseModel(PydBaseModel):
     class Config:
